requirements.txt
selenium
pytest
pytest-html
pytest-xdist
allure-pytest
pyyaml
pytest-rerunfailures
faker
requests 
pytest-httpx
